решена проблема с незакрытием окна по dispose
