+++ src/main/java/gui/GameVisualizer.java
@@ -175,4 +175,8 @@ public class GameVisualizer extends JPanel {
     g.setColor(Color.BLACK);
     drawOval(g, x, y, 5, 5);
   }
+
+  public void stopTimer() {
+    m_timer.cancel();
+  }
 }
